moved all covid-19 related banners into one central file and added documentation
