+++ frontend/lib/letter-of-complaint-splash.tsx
@@ -7,14 +7,7 @@ import { BigList } from './big-list';
 import { OutboundLink } from './google-analytics';
 import { GetStartedButton } from './get-started-button';
 import { OnboardingInfoSignupIntent } from './queries/globalTypes';
-import { Icon } from './icon';
-
-export const MoratoriumWarning = () => (
-<div className="content has-text-centered is-size-7">
-<Icon type="warning" />{' '}Have you been given an eviction notice? <strong>This is illegal.</strong> An Eviction Moratorium is currently in place across New York State. 
-  {' '}<OutboundLink href="https://www.righttocounselnyc.org/moratorium_faq" target="_blank"><strong>Learn more</strong></OutboundLink>
-</div>
-)
+import { MoratoriumWarning } from './covid-banners';
 
 export function LocSplash(): JSX.Element {
   return (
